Move image render functions into pages
Move image page renderes into seperate files
