+++ src/ui/images/constants.rs
@@ -1,0 +1,3 @@
+
+pub const POSSIBLE_EXTENSIONS: [&str; 4] = ["png", "jpg", "ico", "webp"];
+pub const MAX_WIDTH: f32 = 300.;
