+++ src/ui/images/pages/mod.rs
@@ -1,0 +1,12 @@
+mod change_grid_db_id;
+mod shortcut_images_overview;
+mod steam_images_overview;
+mod select_image_type;
+mod pick_new_image;
+
+
+pub use change_grid_db_id::render_page_change_grid_db_id;
+pub use shortcut_images_overview::render_page_shortcut_images_overview;
+pub use steam_images_overview::render_page_steam_images_overview;
+pub use select_image_type::render_page_shortcut_select_image_type;
+pub use pick_new_image::render_page_pick_image;
