+++ src/ui/images/texturestate.rs
@@ -1,0 +1,7 @@
+#[derive(Clone)]
+pub enum TextureDownloadState {
+    Downloading,
+    Downloaded,
+    Loaded(egui::TextureHandle),
+    Failed,
+}
