+++ src/ui/mod.rs
@@ -1,15 +1,14 @@
 mod defines;
 mod ui_backup;
 mod ui_disconnect;
-mod ui_image_download;
 mod ui_import_games;
 mod ui_settings;
 mod uiapp;
+mod images;
 
 pub use defines::*;
 pub use ui_backup::*;
 pub use ui_disconnect::*;
-pub use ui_image_download::*;
 pub use ui_import_games::*;
 pub use ui_settings::*;
 pub use uiapp::*;
